Feat: Implementa o DTO e o Mapper de User no controller e service

+++ src/main/java/lest/dev/RecoBook/User/service/UserService.java
@@ -1,5 +1,7 @@
 package lest.dev.RecoBook.User.service;
 
+import lest.dev.RecoBook.User.dto.UserDTO;
+import lest.dev.RecoBook.User.mapper.UserMapper;
 import lest.dev.RecoBook.User.model.User;
 import lest.dev.RecoBook.User.repository.UserRepository;
 import lombok.AllArgsConstructor;
@@ -12,34 +14,41 @@ import java.util.Optional;
 public class UserService {
 
     private final UserRepository userRepository;
+    private final UserMapper userMapper;
 
-    public User createUser(User user) {
-
-        return userRepository.save(user);
-
-    }
-
-    public Optional<User> detailUser(Long id) {
-
-        return userRepository.findById(id);
-
-    }
-
-    public User updateUser(Long id, User user) {
-
+    public UserDTO detailUser(Long id) {
         Optional<User> userOptional = userRepository.findById(id);
         if (userOptional.isPresent()) {
-            user.setId(id);
-            return userRepository.save(user);
+            User user = userOptional.get();
+            return userMapper.map(user);
         }
         return null;
 
     }
 
-    public void deleteUser(Long id) {
+    public UserDTO createUser(UserDTO userDTO) {
+        User user = userMapper.map(userDTO);
+        return userMapper.map(userRepository.save(user));
+    }
 
-        userRepository.deleteById(id);
+    public UserDTO updateUser(Long id, UserDTO userDTO) {
+        Optional<User> userOptional = userRepository.findById(id);
+        if (userOptional.isPresent()) {
+            userDTO.setId(id);
+            User user = userMapper.map(userDTO);
+            return userMapper.map(userRepository.save(user));
+        }
+        return null;
 
+    }
+
+    public boolean deleteUser(Long id) {
+        Optional<User> user = userRepository.findById(id);
+        if (user.isPresent()) {
+            userRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 
 }
